refactor(langy): one owned chat-engine seam; drop the vestigial casts (ADR-059 Phase 0)
- useLangyChatEngine owns useChat, history hydration, and the engine
  reset; the panel composes two owned seams (engine + recovery) instead
  of reaching into stop/clearError/setMessages internals.
- The messages-as-unknown-as casts into langyThinkingLine and
  deriveWaveActivity were vestigial — UIMessage[] assigns to the
  structural ThinkingMessage[] directly (probed with a scoped tsc).
- The setMessages cast's justification (nested ai version skew) is
  stale — the dep tree dedupes to one ai@6.0.217; what remains is the
  honest wire-boundary cast (verbatim-stored parts re-entering the SDK
  union), stated as such.
- matchesId accepts unknown and narrows once inside, instead of forcing
  an interface->Record cast at every call site.
- The gallery's fixture cast gets its justifying comment (runtime-built
  tool parts cannot satisfy the SDK's per-state discriminated union).
